- Separate components into individual files - Add Bluehawk markup
