+++ examples/react-native/v12/TestApp/src/components/subscribeApi/Unsubscribe.tsx
@@ -1,0 +1,58 @@
+// :snippet-start: unsubscribe-query-full
+import React, {useEffect, useState} from 'react';
+import {useRealm, useQuery} from '@realm/react';
+import {View, Text, Button} from 'react-native';
+
+import {Bird} from './models/Bird';
+import {Subscription} from 'realm/dist/bundle';
+
+export const Unsubscribe = () => {
+  const realm = useRealm();
+  const birds = useQuery(Bird);
+  // :remove-start:
+  const [birdsSubscription, setBirdsSubscription] =
+    useState<Subscription | null>();
+
+  useEffect(() => {
+    const getSubscription = async () => {
+      // Get an initial subscription named "Initial birds".
+      const subscription = realm.subscriptions.findByName('Initial birds');
+
+      // Set subscription to stateful variable or manage
+      // it in `useEffect`.
+      setBirdsSubscription(subscription);
+    };
+
+    getSubscription().catch(console.error);
+  }, []);
+  // :remove-end:
+
+  const unsubscribeFromQuery = () => {
+    birds.unsubscribe();
+    setBirdsSubscription(realm.subscriptions.findByName('Initial birds')); // :remove:
+  };
+
+  // :replace-start: {
+  //    "terms": {
+  //       '"testID=unsubscribe"': ""
+  //    }
+  // }
+  return (
+    <View>
+      {/* :remove-start: */}
+      <Text testID="bird-subscription">
+        Subscription name: {birdsSubscription?.name}
+      </Text>
+      {/* :remove-end: */}
+      <Button
+        testID="unsubscribe"
+        title="Unsubscribe"
+        onPress={() => {
+          unsubscribeFromQuery();
+        }}
+      />
+    </View>
+  );
+  // :replace-end:
+};
+// :snippet-end:
